[feat] #44  디테일 뷰 리팩토링
extension으로 작성한 코드들을 struct로 변경
